removed old main, renamed new main to main
