+++ src/main.py
@@ -3,7 +3,7 @@ from PySide6.QtCore import (QCoreApplication, QMetaObject, QObject, QPoint,
     QRect, QSize, QUrl, Qt, QTimer)
 from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor, QFont,
     QFontDatabase, QIcon, QLinearGradient, QPalette, QPainter, QPixmap, QPen,
-    QRadialGradient)
+    QRadialGradient, QAction, QKeySequence)
 from PySide6.QtWidgets import *
 from PySide6 import QtWidgets
 from utils import *
@@ -12,93 +12,508 @@ from classes import *
 import json
 
 
+# because things are often strings, we need to hard disallow
+# some names to prevent things from breaking
+# all lowercase so we can check `name.lower() in DISALLOWED_NAMES`
+DISALLOWED_NAMES = set(
+    ["no winner",
+     "delayed",
+     "select winner..."]
+)
+
 
 class MainWindow(QtWidgets.QMainWindow):
     players: list[Player] = []
     rounds: list[Round] = []
 
-    def __init__(self):
+    def __init__(self, launch_data):
         QtWidgets.QMainWindow.__init__(self)
         self.ui = Ui_MainWindow()
         self.ui.setupUi(self)
+        self.setWindowTitle("Swiss Bracket Maker")
 
         self.ui.ImportPlayersFileButton.clicked.connect(self.import_players_from_file)
         self.ui.importPlayersClipboardButton.clicked.connect(self.import_players_from_clipboard)
         self.ui.generateRound.clicked.connect(self.generate_round)
         self.ui.generateBracket.clicked.connect(self.on_generate_final_bracket_clicked)
-        self.ui.importButton.clicked.connect(self.import_session)
         self.ui.exportButton.clicked.connect(self.export_session)
 
         # set column headers in players table
         self.ui.playersTableWidget.setColumnCount(5)
         self.ui.playersTableWidget.setHorizontalHeaderLabels(["Drop", "Name", "Score", "Resistance", "Win Percentage"])
-        self.ui.playersTableWidget.cellChanged.connect(self.on_player_cell_changed)
+        # self.ui.playersTableWidget.cellChanged.connect(self.on_player_cell_changed)
+        self.ui.tabWidget.currentChanged.connect(self.tab_change_controller)
 
-    def import_session(self):
-        file_name, _ = QFileDialog.getOpenFileName(
-            self,
-            "Load Tournament Data",
-            "",
-            "JSON Files (*.json)"
-        )
-        if not file_name:
+        # Check if we loaded from file or not
+        print(f"Launching with previous data: {launch_data is not None}")
+        if launch_data:
+            self.import_session(launch_data)
+
+    def tab_change_controller(self, index):
+        """
+        We recalculate the players table every time we click on the tab
+        """
+        tabname = self.ui.tabWidget.tabText(index)
+        print("Current tab name:", tabname)
+
+        if tabname == "Players":
+            self.create_players_table()
+
+
+    def import_players_from_file(self):
+        filename, _ = QFileDialog.getOpenFileName()
+        if not filename:
             return
-        # Parse the file
-        try:
-            with open(file_name, "r", encoding="utf-8") as f:
-                data = json.load(f)
-        except Exception as e:
-            QMessageBox.critical(self, "Import Failed", f"Could not load file:\n{e}")
-            return None, None
+        count = 0
 
-        # Step 1: Rebuild players
-        self.players = []
-        for p_data in data.get("players", []):
-            p = Player(p_data["name"])
-            p.clean_name = p_data["clean_name"]
-            p.score = p_data["score"]
-            p.resistance = p_data["resistance"]
-            p.dropped = p_data["dropped"]
-            p.winpercentage = p_data["winpercentage"]
-            self.players.append(p)
-            self.create_player_table_entry(p)
+        current_player_names = {player.name.strip().lower() for player in self.players}
+        with open(filename, 'r') as f:
+            for name in f:
+                name = name.strip()
+                if not name or name.lower() in current_player_names:
+                    print(f"Duplicate name {name}")
+                    continue
+                if name.lower() in DISALLOWED_NAMES:
+                    print(f"Disallowed name {name}")
+                    continue
+                new_player = Player(name)
+                self.players.append(new_player)
+                count += 1
+                current_player_names.add(name.lower())
+        self.ui.settingsMessage.setText(f"Imported {count} players successfully")
 
-        # Step 2: Rebuild rounds and matchups
-        for r_data in data.get("rounds", []):
-            matchups = []
-            for m_data in r_data["matchups"]:
-                p1 = get_player_by_name(self.players,m_data["player1"])
-                p2 = get_player_by_name(self.players,m_data["player2"]) if m_data["player2"] else None
 
-                matchup = Matchup(p1, p2, m_data["notes"])
-                matchup.score_player1 = m_data["score_player1"]
-                matchup.score_player2 = m_data["score_player2"]
+    def import_players_from_clipboard(self):
+        data = get_clipboard_data()
+        names = data.split('\n')
+        count = 0
 
-                # Restore winner if set
-                if m_data["winner"]:
-                    matchup.winner = get_player_by_name(self.players,m_data["winner"])
+        current_player_names = {player.name.strip().lower() for player in self.players}
+        for name in names:
+            name = name.strip()
+            if not name or name.lower() in current_player_names:
+                    print(f"Duplicate name {name}")
+                    continue
+            if name.lower() in DISALLOWED_NAMES:
+                print(f"Disallowed name {name}")
+                continue
+            new_player = Player(name)
+            self.players.append(new_player)
+            count += 1
+            current_player_names.add(name.lower())
+        self.ui.settingsMessage.setText(f"Imported {count} players successfully")
 
-                matchups.append(matchup)
-            self.generate_round(matchups=matchups, locked=r_data["locked"])
+    def create_players_table(self):
+        # First calculate all the players' stats
+        player_info_list = calculate_players_stats(self.players, self.rounds)
 
-        # Step 3: Restore player.matches (match history)
-        for r in self.rounds:
-            for m in r.matchups:
-                if m.player1:
-                    m.player1.matches.append(m)
-                if m.player2:
-                    m.player2.matches.append(m)
+        print([(p.player.name, p.score, p.resistance) for p in player_info_list if p.score > 10])
 
-        QMessageBox.information(self, "Import Successful", f"Tournament loaded from:\n{file_name}")
+        # Clear the table
+        self.ui.playersTableWidget.setSortingEnabled(False)
+        self.ui.playersTableWidget.clearContents()
+        self.ui.playersTableWidget.setRowCount(0)
 
-        # Make all the tabs
-        pass
+        # Create the table
+        for player_info in player_info_list:
+            self.create_player_table_entry(player_info)
+
+        self.ui.playersTableWidget.setSortingEnabled(True)
+
+
+    def create_player_table_entry(self, player_info: PlayerInfo):
+        # Get the table row index
+        rowPosition = self.ui.playersTableWidget.rowCount()
+        self.ui.playersTableWidget.insertRow(rowPosition)
+
+        # Create a centered checkbox cell
+        checkbox = QCheckBox()
+        checkbox.setChecked(player_info.player.dropped)
+        checkbox.stateChanged.connect(lambda state, player=player_info.player: self.on_checkbox_state_changed(state, player))
+        # Create a widget to hold the layout
+        centered_widget = QWidget()
+        layout = QHBoxLayout(centered_widget)
+        layout.addWidget(checkbox)
+        layout.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        layout.setContentsMargins(0, 0, 0, 0)
+
+        # Dropped checkbox
+        self.ui.playersTableWidget.setCellWidget(rowPosition, 0, centered_widget)
+        # Player name
+        self.ui.playersTableWidget.setItem(rowPosition, 1, QTableWidgetItem(player_info.player.name))
+        # Score
+        player_score = QTableWidgetItem()
+        player_score.setData(Qt.ItemDataRole.EditRole, player_info.score)
+        self.ui.playersTableWidget.setItem(rowPosition, 2, player_score)
+        # TODO: check if double convert is needed with current python version and floats https://stackoverflow.com/questions/455612/limiting-floats-to-two-decimal-points
+        # TODO: number styling to be consistent? 0 padding etc
+        # Resistance
+        player_res = QTableWidgetItem()
+        player_res.setData(Qt.ItemDataRole.EditRole, round(player_info.resistance, 2))
+        player_res.setData(Qt.ItemDataRole.DisplayRole, f"{round(player_info.resistance, 2)}")
+        self.ui.playersTableWidget.setItem(rowPosition, 3, player_res)
+        # self.ui.playersTableWidget.setItem(rowPosition, 3, QTableWidgetItem("{:.2f}%".format(player_info.resistance)))
+
+        # Win percentage
+        if player_info.n_played:
+            player_win_percentage = player_info.n_wins / player_info.n_played * 100
+        else:
+            player_win_percentage = 0
+
+        # TODO: make it sort numerically instead of by string (i.e. 100 > 50)
+        player_win = QTableWidgetItem()
+        player_win.setData(Qt.ItemDataRole.EditRole, round(player_win_percentage, 2))
+        player_win.setData(Qt.ItemDataRole.DisplayRole, f"{round(player_win_percentage, 2)}%")
+        self.ui.playersTableWidget.setItem(rowPosition, 4, player_win)
+        # self.ui.playersTableWidget.setItem(rowPosition, 4, QTableWidgetItem("{:.2f}%".format(player_win_percentage)))
+
+
+
+    def generate_round(self):
+        if len(self.players) == 0:
+            self.ui.settingsMessage.setText(f"Import players before generating a round!")
+            return
+
+        if not self.confirm_start_round_generation():
+            return
+
+        # Generate the matchups and display them
+        matchups = generate_matchups(self.players, self.rounds)
+
+        new_round = Round(matchups)
+        self.rounds.append(new_round)
+        round_number = len(self.rounds)
+        for matchup in new_round.matchups:
+            if not matchup.player2:
+                matchup.winner = matchup.player1
+                matchup.score_player1 = 1.0
+                matchup.notes = "BYE"
+
+        self.generate_round_tab(new_round, round_number)
+
+
+    def generate_round_tab(self, round: Round, round_number: int):
+        # Create the table for the new round
+        table = QTableWidget()
+        table.setColumnCount(6)
+        table.setHorizontalHeaderLabels(["P1", "P2", "Winner", "P1Score", "P2Score", "Notes"])
+        table.setSortingEnabled(True)
+
+        for idx, matchup in enumerate(round.matchups):
+            table.insertRow(idx)
+            winner_combo = QComboBox()
+            winner_combo.setEditable(True)
+            winner_combo.lineEdit().setReadOnly(True)
+
+            winner_combo.lineEdit().setPlaceholderText("Select winner...")
+            winner_combo.setStyleSheet("QComboBox { combobox-popup: 0; }")  # Optional styling
+
+            winner_combo.addItem(matchup.player1)
+            if matchup.player2 != "":
+                winner_combo.addItem(matchup.player2)
+            winner_combo.addItem("No Winner")
+            winner_combo.addItem("Delayed")
+
+            # If a winner already exists (e.g., for BYEs), select them
+            if matchup.winner:
+                winner_combo.setCurrentText(matchup.winner)
+            else:
+                winner_combo.setCurrentIndex(-1)  # Show placeholder, no selection
+
+            winner_combo.setProperty("matchup", matchup)
+
+            # Name columns should not be editable
+            p1_item = QTableWidgetItem(matchup.player1)
+            p2_item = QTableWidgetItem(matchup.player2)
+            p1_item.setFlags(p1_item.flags() & ~Qt.ItemIsEditable)
+            p2_item.setFlags(p2_item.flags() & ~Qt.ItemIsEditable)
+            table.setItem(idx, 0, p1_item)
+            table.setItem(idx, 1, p2_item)
+
+            table.setCellWidget(idx, 2, winner_combo)
+            winner_combo.currentTextChanged.connect(lambda _, combo=winner_combo, table=table: self.on_winner_changed(combo, table))
+
+            p1_score_item = QTableWidgetItem(str(matchup.score_player1))
+            table.setItem(idx, 3, p1_score_item)
+
+            p2_score_item = QTableWidgetItem(str(matchup.score_player2))
+            table.setItem(idx, 4, p2_score_item)
+
+            notes_item = QTableWidgetItem(str(matchup.notes))
+
+            table.setItem(idx, 5, notes_item)
+
+            # attach round index and matchup data to the first column
+            p1_item.setData(Qt.UserRole, {"round_idx": round_number-1, "matchup": matchup})
+
+
+        table.cellChanged.connect(self.on_cell_changed)
+
+
+        container = QWidget()
+        layout = QVBoxLayout(container)
+
+        button_row = QHBoxLayout()
+        button_row.setSpacing(10)
+
+        paste_winners_button = QPushButton("Paste Winners")
+        paste_winners_button.clicked.connect(lambda: self.paste_winners(table))
+        button_row.addWidget(paste_winners_button)
+
+        fill_unfilled_button = QPushButton("Set \"No Winner\" for Unfilled")
+        fill_unfilled_button.clicked.connect(lambda: self.unfilled_to_no_winner(table))
+        button_row.addWidget(fill_unfilled_button)
+
+        clipboard_button = QPushButton("Round to Clipboard")
+        clipboard_button.clicked.connect(lambda: self.round_to_clipboard(table))
+        button_row.addWidget(clipboard_button)
+
+        # TODO: rethink; remove for now, possibly just remove or only for last round or error if not last round or something
+        # delete_button = QPushButton("Delete Round")
+        # delete_button.setStyleSheet("background-color: lightcoral;")  # visually distinct
+        # delete_button.clicked.connect(lambda _, t=table, r=round, i=round_number-1: self.confirm_delete_round(t, r, i))
+        # button_row.addWidget(delete_button)
+
+        layout.addLayout(button_row)
+
+        layout.addWidget(table)
+        container.setLayout(layout)
+
+        # Create the tab
+        self.ui.tabWidget.addTab(container, f"R{round_number}")
+        self.ui.settingsMessage.setText(f"Created round {round_number}.")
+
+
+    def confirm_start_round_generation(self):
+        """
+        Displays some (possibly worrying) information about the rounds so far
+        and asks for confirmation that the next round is to be generated
+        """
+
+        # if first round, don't be scared
+        if not self.rounds:
+            return True
+
+        # otherwise we do some sanity checks
+        # first we check if there's delays left from very old rounds
+        text = ""
+        for i, old_round in enumerate(self.rounds[:-1]):
+            delays = 0
+            for matchup in old_round.matchups:
+                if matchup.winner == "Delayed":
+                    delays += 1
+            if delays:
+                text += f"Round {i+1} still has {delays} delayed matches.\n"
+
+        # then we gather some numbers from the last round
+        last_round = self.rounds[-1]
+        non_binary_score = False
+        delays = 0
+        no_winners = 0
+        winners = 0
+        for matchup in last_round.matchups:
+            if not matchup.winner:
+                warning_text = f"Matchup {matchup.player1} vs {matchup.player2} has no winner selected!\n"
+                warning_text += "If this is intentional, either select \"No Winner\" or \"Delayed\" before generating the next round."
+                QMessageBox.warning(self, "Incomplete information", warning_text)
+                return False
+            elif matchup.winner == "Delayed":
+                delays += 1
+            elif matchup.winner == "No Winner":
+                no_winners += 1
+            else:
+                winners += 1
+            if matchup.score_player1 + matchup.score_player2 not in [0,1]:
+                non_binary_score = (matchup.score_player1, matchup.score_player2)
+        if non_binary_score:
+            text += f"Last round has matches with nonstandard scores, for example {non_binary_score}\n"
+        text += f"Last round had {winners} winners, {no_winners} matches without winners and {delays} delayed matches still to be played.\n"
+        if delays:
+            text += "Delayed matches will count as half a point to each player for the purposes of generating the next round."
+        text += "Are you sure you want to generate the next round?"
+
+        reply = QMessageBox.question(
+            self,
+            "Confirm Round Generation",
+            text,
+            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No
+        )
+
+        if reply == QMessageBox.StandardButton.Yes:
+            return True
+        else:
+            return False
+
+    def on_winner_changed(self, combo: QComboBox, table: QTableWidget):
+        matchup = combo.property("matchup")
+
+        winner_name = combo.currentText()
+        matchup.winner = winner_name
+        print(f"{matchup} winner changed to {matchup.winner}")
+        if winner_name == matchup.player1:
+            matchup.score_player1 = 1.0
+            matchup.score_player2 = 0.0
+        elif winner_name == matchup.player2:
+            matchup.score_player1 = 0.0
+            matchup.score_player2 = 1.0
+        else:  # No Winner / Delayed
+            matchup.score_player1 = 0.0
+            matchup.score_player2 = 0.0
+
+        self.update_matchup_row_scores(table, matchup)
+
+    def on_cell_changed(self, row, col):
+        table = self.sender()
+        item = table.item(row, col)
+        # first column stores the data in userrole
+        matchup = table.item(row, 0).data(Qt.UserRole)["matchup"]
+        if not matchup:
+            print(f"Error: item at row {row} and col {col} does not have matchup data attached")
+            return
+
+        value = item.text()
+
+        if col == 3:
+            matchup.score_player1 = float(value)
+            print(f"Updated p1 score in {matchup} to {value}")
+        elif col == 4:
+            matchup.score_player2 = float(value)
+            print(f"Updated p2 score in {matchup} to {value}")
+        elif col == 5:
+            matchup.notes = value
+            print(f"Updated notes in {matchup} to {value}")
+
+
+    def update_matchup_row_scores(self, table: QTableWidget, matchup: Matchup):
+        """
+        Updates *the view of* the scores for a single row after updating the winner.
+        Is O(N), if that becomes problematic caching a mapping dict can make it essentially
+        O(1) but is a bit painful.
+        """
+        row = self.find_round_row_by_matchup(table, matchup)
+        if row == -1:
+            print("Error: matchup row to be updated not found")
+            return
+
+        # stops `on_cell_changed` from being hit here
+        table.blockSignals(True)
+        table.item(row, 3).setText(str(matchup.score_player1))
+        table.item(row, 4).setText(str(matchup.score_player2))
+        table.blockSignals(False)
+
+    def find_round_row_by_matchup(self, table: QTableWidget, matchup: Matchup):
+        for row in range(table.rowCount()):
+            p1_item = table.item(row, 0)  # first column has the data in UserRole
+            if p1_item.data(Qt.UserRole)["matchup"] is matchup:
+                return row
+        return -1
+
+
+    def paste_winners(self, table):
+        # Get clipboard text and split into lines/names
+        text = get_clipboard_data()
+        winners = [name.strip().lower() for name in text.splitlines() if name.strip()]
+
+        # Go through each row of the table and find matches
+        for row in range(table.rowCount()):
+            player1_item = table.item(row, 0)
+            player2_item = table.item(row, 1)
+
+            player1 = player1_item.text().strip().lower()
+            player2 = player2_item.text().strip().lower()
+
+            # names to players
+            for winner in winners:
+                if winner == player1 or winner == player2:
+                    widget = table.cellWidget(row, 2)
+                    if isinstance(widget, QComboBox):
+                        index = widget.findText(winner, Qt.MatchFixedString)
+                        if index >= 0:
+                            widget.setCurrentIndex(index)
+                    break
+
+    def unfilled_to_no_winner(self, table):
+        """
+        Fills every matchup with no winner selected to `No Winner`.
+        Nothing selected usually means "not played yet", which
+        blocks round generation. `No Winner`, on the other hand, means
+        the match will not be played, and will count as a loss for both players.
+        """
+        # first we count for a popup
+        count = 0
+        for row in range(table.rowCount()):
+            matchup: Matchup = table.item(row, 0).data(Qt.UserRole)["matchup"]
+            if not matchup.winner:
+                count += 1
+
+        reply = QMessageBox.question(
+            self,
+            "Confirm \"No Winner\" fill",
+            f"{count} matchups will be set to \"No Winner\". Continue?",
+            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No
+        )
+
+        if reply == QMessageBox.StandardButton.Yes:
+            for row in range(table.rowCount()):
+                matchup: Matchup = table.item(row, 0).data(Qt.UserRole)["matchup"]
+                if not matchup.winner:
+                    widget = table.cellWidget(row, 2)
+                    if isinstance(widget, QComboBox):
+                        index = widget.findText("No Winner", Qt.MatchFixedString)
+                        if index >= 0:
+                            widget.setCurrentIndex(index)
+
+
+    def round_to_clipboard(self, table: QTableWidget):
+        round_index = table.item(0, 0).data(Qt.UserRole)["round_idx"]
+        print(f"Saving round {round_index+1} to clipboard")
+        round = self.rounds[round_index]
+        # take into account only previous rounds to get stats pre-round
+        player_stats_dict = calculate_players_stats(self.players, self.rounds[:round_index], as_dict=True)
+
+        output_str = ""
+        for row in range(table.rowCount()):
+            matchup: Matchup = table.item(row, 0).data(Qt.UserRole)["matchup"]
+            output_str += self.format_match_for_clipboard(matchup, player_stats_dict) + "\n"
+
+        QApplication.clipboard().setText(output_str)
+        print("Copied")
+
+
+    def format_match_for_clipboard(self, matchup: Matchup, player_stats_dict: dict[str, PlayerInfo]) -> str:
+        """
+        Just some annoying formatting to turn a matchup into
+        `one_name (6.0/1) — other_name (7.0)`
+        where 6.0 and 7.0 are their scores, and the /1 indicates one
+        currently delayed game for the first player.
+        """
+
+        stats1 = player_stats_dict[matchup.player1]
+        result = matchup.player1
+        result += f" ({stats1.score}"
+        if stats1.active_delays:
+            result += f"/{stats1.active_delays}"
+        result += ") — "
+        if not matchup.player2:
+            result += "BYE"
+        else:
+            stats2 = player_stats_dict[matchup.player2]
+            result += matchup.player2
+            result += f" ({stats2.score}"
+            if stats2.active_delays:
+                result += f"/{stats2.active_delays}"
+            result += ")"
+
+        return result
 
 
     def export_session(self):
         # Convert data to dict
+        player_dump = {
+            player.name: player.dropped for player in self.players
+        }
         data = {
-            "players": [p.to_dict() for p in self.players],
+            "players": player_dump,
             "rounds": [r.to_dict() for r in self.rounds]
         }
 
@@ -125,409 +540,58 @@ class MainWindow(QtWidgets.QMainWindow):
             QMessageBox.critical(self, "Export Failed", f"Could not save file:\n{e}")
 
 
-    def on_player_cell_changed(self, row, column):
-        # Get Player object from Name column
-        player_item = self.ui.playersTableWidget.item(row, 1).text()
-        if player_item is None:
-            return
+    # TODO: if we dont want the import button to exist this doesn't need to remove tabs and such
+    def import_session(self, data):
+        # Delete old tabs
+        tabs_to_remove = []
+        for i in range(self.ui.tabWidget.count())[::-1]:
+            tab_text = self.ui.tabWidget.tabText(i)
+            if 'R' in tab_text:
+                tabs_to_remove.append(i)
+        for i in sorted(tabs_to_remove, reverse=True):
+            self.ui.tabWidget.removeTab(i)
 
-        player = [x for x in self.players if x.name == player_item][0]
-        if not player:
-            return
+        print("Rebuilding session")
+        start = time.time()
 
-        if column == 2:
-            try:
-                print(f"Updated {player.name}: Score from {player.score} to {float(self.ui.playersTableWidget.item(row, column).text())}")
-                player.score = float(self.ui.playersTableWidget.item(row, column).text())
-            except ValueError:
-                player.score = 0.0
-        elif column == 3:
-            try:
-                print(f"Updated {player.name}: Resistance from {player.resistance} to {float(self.ui.playersTableWidget.item(row, column).text())}")
-                player.resistance = float(self.ui.playersTableWidget.item(row, column).text())
-            except ValueError:
-                player.resistance = 0.0
+        # Step 1: Rebuild players
+        self.players = []
+        for p_name, p_dropped in data.get("players", {}).items():
+            p = Player(p_name, dropped=p_dropped)
+            self.players.append(p)
 
+        # Step 2: Rebuild rounds and matchups
+        self.rounds = []
+        for round_number, r_data in enumerate(data.get("rounds", [])):
+            matchups = []
+            for saved_matchup in r_data["matchups"]:
+                p1 = saved_matchup["player1"]
+                p2 = saved_matchup["player2"]
+                notes = saved_matchup["notes"]
 
-    def import_players_from_file(self):
-        filename, _ = QFileDialog.getOpenFileName()
-        if not filename:
-            return
-        count = 0
+                matchup = Matchup(p1, p2, notes)
+                matchup.score_player1 = saved_matchup["score_player1"]
+                matchup.score_player2 = saved_matchup["score_player2"]
+                matchup.winner = saved_matchup["winner"]
 
-        current_player_names = {player.name.strip().lower() for player in self.players}
-        with open(filename, 'r') as f:
-            for name in f:
-                name = name.strip()
-                if not name or name.lower() in current_player_names:
-                    print(f"Duplicate name {name}")
-                    continue
-                new_player = Player(name)
-                self.players.append(new_player)
-                self.create_player_table_entry(new_player)
-                count += 1
-                current_player_names.add(name.lower())
-        self.ui.settingsMessage.setText(f"Imported {count} players successfully")
+                matchups.append(matchup)
+            saved_round = Round(matchups)
+            self.rounds.append(saved_round)
+            self.generate_round_tab(saved_round, round_number + 1)
 
-
-    def import_players_from_clipboard(self):
-        data = get_clipboard_data()
-        names = data.split('\n')
-        count = 0
-
-        current_player_names = {player.name.strip().lower() for player in self.players}
-        for name in names:
-            name = name.strip()
-            if not name or name.lower() in current_player_names:
-                    print(f"Duplicate name {name}")
-                    continue
-            new_player = Player(name)
-            self.players.append(new_player)
-            self.create_player_table_entry(new_player)
-            count += 1
-            current_player_names.add(name.lower())
-        self.ui.settingsMessage.setText(f"Imported {count} players successfully")
-
-
-    def create_player_table_entry(self, player: Player):
-        # Get the table row index
-        rowPosition = self.ui.playersTableWidget.rowCount()
-        self.ui.playersTableWidget.insertRow(rowPosition)
-        self.ui.playersTableWidget.setSortingEnabled(True)
-
-        # Create a centered checkbox cell
-        checkbox = QCheckBox()
-        checkbox.stateChanged.connect(lambda state, p=player: self.on_checkbox_state_changed(state, p))
-        # Create a widget to hold the layout
-        centered_widget = QWidget()
-        layout = QHBoxLayout(centered_widget)
-        layout.addWidget(checkbox)
-        layout.setAlignment(Qt.AlignmentFlag.AlignCenter)
-        layout.setContentsMargins(0, 0, 0, 0)
-
-        # Set the widget in the cell
-        self.ui.playersTableWidget.setCellWidget(rowPosition, 0, centered_widget)
-
-        self.ui.playersTableWidget.setItem(rowPosition, 1, QTableWidgetItem(player.name))
-        self.ui.playersTableWidget.setItem(rowPosition, 2, QTableWidgetItem(str(player.score)))
-        self.ui.playersTableWidget.setItem(rowPosition, 3, QTableWidgetItem(str(player.resistance)))
-        self.ui.playersTableWidget.setItem(rowPosition, 4, QTableWidgetItem(str(player.winpercentage) + "%"))
+        print(f"Session rebuilding took {time.time() - start} seconds")
+        # Make all the tabs
+        pass
 
 
     def on_checkbox_state_changed(self, state: int, player: Player):
-            if state == Qt.CheckState.Checked.value:
-                print(f"{player.name} is now dropped.")
-                player.dropped = True
-            else:
-                print(f"{player.name} is now active.")
-                player.dropped = False
+        if state == Qt.CheckState.Checked.value:
+            print(f"{player.name} is now dropped.")
+            player.dropped = True
+        else:
+            print(f"{player.name} is now active.")
+            player.dropped = False
 
-
-    def generate_round(self, checked: bool = False, matchups=[], locked=False):
-        if len(self.players) == 0:
-            self.ui.settingsMessage.setText(f"Import players before generating a round!")
-            return
-        # Check if last round was locked
-        if len(self.rounds) >=1 and not self.rounds[-1].locked:
-            self.ui.settingsMessage.setText("Previous round is not locked.")
-            return
-
-        # Create the table for the new round
-        table = QTableWidget()
-        table.setColumnCount(6)
-        table.setHorizontalHeaderLabels(["P1", "P2", "Winner", "Notes", "P1Score", "P2Score"])
-        table.setSortingEnabled(True)
-
-        # Generate the matchups and display them
-        if not matchups:
-            matchups = generate_matchups(self.players)
-        new_round = Round(matchups, locked)
-        self.rounds.append(new_round)
-        round_number = len(self.rounds)
-        for idx, matchup in enumerate(matchups):
-            table.insertRow(idx)
-            # in case of BYE
-            winner = Player("")
-            p1 = matchup.player1
-            if matchup.player2:
-                p2 = matchup.player2
-            else:
-                p2 = Player("-")
-                winner = p1
-                matchup.score_player1 = 1.0
-
-
-            table.setItem(idx, 0, QTableWidgetItem(str(p1.name)))
-            table.setItem(idx, 1, QTableWidgetItem(str(p2.name)))
-
-            combo = QComboBox()
-            combo.setEditable(True)
-            combo.lineEdit().setReadOnly(True)
-
-            combo.lineEdit().setPlaceholderText("Select winner...")
-            combo.setStyleSheet("QComboBox { combobox-popup: 0; }")  # Optional styling
-
-            combo.addItem(p1.name)
-            if p2.name != "-":
-                combo.addItem(p2.name)
-
-            # If a winner already exists (e.g., for BYEs), select them
-            if winner.name:
-                combo.setCurrentText(winner.name)
-            else:
-                combo.setCurrentIndex(-1)  # Show placeholder, no selection
-
-            combo.currentTextChanged.connect(lambda name, row=idx: self.on_winner_changed(row, name, table, matchups))
-            table.setCellWidget(idx, 2, combo)
-
-            table.setItem(idx, 3, QTableWidgetItem(str(matchup.notes)))
-            table.setItem(idx, 4, QTableWidgetItem(str(matchup.score_player1)))
-            table.setItem(idx, 5, QTableWidgetItem(str(matchup.score_player2)))
-            if matchup.winner:
-                index = combo.findText(matchup.winner.name)
-                combo.setCurrentIndex(index)
-                self.on_winner_changed(idx, matchup.winner.name, table, matchups)
-
-        table.cellChanged.connect(lambda row, col, t=table, m=matchups: self.on_cell_changed(t, m, row, col))
-
-        container = QWidget()
-        layout = QVBoxLayout(container)
-
-        button_row = QHBoxLayout()
-        button_row.setSpacing(10)
-
-        lock_button = QPushButton("Lock Round")
-        lock_button.clicked.connect(lambda _, t=table, r=new_round: self.lock_round(t, r))
-        button_row.addWidget(lock_button)
-
-        paste_winners_button = QPushButton("Paste Winners")
-        paste_winners_button.clicked.connect(lambda: self.paste_winners(table))
-        button_row.addWidget(paste_winners_button)
-
-        delete_button = QPushButton("Delete Round")
-        delete_button.setStyleSheet("background-color: lightcoral;")  # visually distinct
-        delete_button.clicked.connect(lambda _, t=table, r=new_round, i=round_number-1: self.confirm_delete_round(t, r, i))
-        button_row.addWidget(delete_button)
-
-        layout.addLayout(button_row)
-
-        layout.addWidget(table)
-        container.setLayout(layout)
-
-        # Create the tab
-        self.ui.tabWidget.addTab(container, f"R{round_number}")
-        self.ui.settingsMessage.setText(f"Created round {round_number}.")
-
-
-    def confirm_delete_round(self, table: QTableWidget, round: Round, round_index: int):
-        reply = QMessageBox.question(
-            self,
-            "Confirm Delete Round",
-            f"Are you sure you want to delete Round {round_index + 1}?\n"
-            "This will disable editing and undo any scores if the round was locked.",
-            QMessageBox.StandardButton.Yes | QMessageBox.StandardButton.No
-        )
-
-        if reply == QMessageBox.StandardButton.Yes:
-            self.delete_round(table, round, round_index)
-
-
-    def delete_round(self, table: QTableWidget, round: Round, round_index: int):
-        if round.locked:
-            # Undo scores
-            for matchup in round.matchups:
-                matchup.player1.score -= matchup.score_player1
-                if matchup.player2:
-                    matchup.player2.score -= matchup.score_player2
-
-            print(f"Reverted scores for round {round_index + 1}")
-
-        # Mark round as deleted
-        round.deleted = True
-        round.locked = True  # Treat as locked for future checks
-
-        # Disable all editing in table
-        for row in range(table.rowCount()):
-            for col in range(table.columnCount()):
-                item = table.item(row, col)
-                if item:
-                    item.setFlags(item.flags() & ~Qt.ItemIsEditable)
-            # Disable dropdowns
-            widget = table.cellWidget(row, 2)
-            if isinstance(widget, QComboBox):
-                widget.setEnabled(False)
-
-        # Grey out table (optional styling)
-        table.setStyleSheet("color: gray;")
-
-        # Rename tab
-        tab_index = round_index + 3
-        tab_text = self.ui.tabWidget.tabText(tab_index)
-        self.ui.tabWidget.setTabText(tab_index, f"DELETED ({tab_text})")
-
-        # Refresh player table
-        self.update_player_table()
-
-    def on_winner_changed(self, row, winner_name, table, matchups):
-        if not winner_name.strip():
-            print(f"No winner selected for row {row}")
-            return  # Do not update scores if blank selected
-
-        winner = get_player_by_name(self.players, winner_name)
-        matchup = matchups[row]
-        matchup.winner = winner
-
-        if winner == matchup.player1:
-            matchup.score_player1 = 1.0
-            matchup.score_player2 = 0.0
-        elif winner == matchup.player2:
-            matchup.score_player2 = 1.0
-            matchup.score_player1 = 0.0
-
-        table.setItem(row, 4, QTableWidgetItem(str(matchup.score_player1)))
-        table.setItem(row, 5, QTableWidgetItem(str(matchup.score_player2)))
-
-        print(f"Winner for matchup {row} set to {winner.name}")
-
-
-    def paste_winners(self, table):
-        # Get clipboard text and split into lines/names
-        text = get_clipboard_data()
-        winners = [name.strip().lower() for name in text.splitlines() if name.strip()]
-
-        # Go through each row of the table and find matches
-        for row in range(table.rowCount()):
-            player1_item = table.item(row, 0)
-            player2_item = table.item(row, 1)
-
-            if not player1_item or not player2_item:
-                continue  # Skip if row is incomplete
-
-            player1 = player1_item.text().strip().lower()
-            player2 = player2_item.text().strip().lower()
-
-            # names to players
-            for winner in winners:
-                if winner == player1 or winner == player2:
-                    widget = table.cellWidget(row, 2)
-                    if isinstance(widget, QComboBox):
-                        index = widget.findText(winner, Qt.MatchFixedString)
-                        if index >= 0:
-                            widget.setCurrentIndex(index)
-                    break  # Move to next row after a match
-
-
-    def on_cell_changed(self, table: QTableWidget, matchups: list, row: int, col: int):
-        if col == 0:
-            new_value = table.item(row, col).text()
-            new_player = get_player_by_name(self.players, new_value)
-            matchups[row].player1 = new_player
-            print(f"Updated player1 for matchup {row} to {new_player.name}")
-        elif col == 1:
-            new_value = table.item(row, col).text()
-            new_player = get_player_by_name(self.players, new_value)
-            if new_player:
-                matchups[row].player2 = new_player
-                print(f"Updated player2 for matchup {row} to {new_player.name}")
-        elif col == 2:
-            winner_name = table.item(row, col).text()
-            winner = get_player_by_name(self.players, winner_name)
-            matchups[row].winner = winner
-            matchup = get_matchup_by_player(matchups, winner)
-            if winner == matchup.player1:
-                matchup.score_player1 = 1.0
-                table.item(row, col+2).setText("1.0")
-            elif winner == matchup.player2:
-                matchup.score_player2 = 1.0
-                table.item(row, col+3).setText("1.0")
-            print(f"Updated winner for matchup {row} to {winner.name}")
-        elif col == 3:
-            new_value = table.item(row, col).text()
-            matchups[row].notes = new_value
-            print(f"Updated notes for matchup {row} to {new_value}")
-        elif col == 4:
-            new_value = table.item(row, col).text()
-            matchups[row].score_player1 = float(new_value)
-            print(f"Updated p1score matchup {row} to {new_value}")
-        elif col == 5:
-            new_value = table.item(row, col).text()
-            matchups[row].score_player2 = float(new_value)
-            print(f"Updated p2score matchup {row} to {new_value}")
-
-
-    def lock_round(self, table: QTableWidget, round: Round):
-        if not round.locked:
-            # Disable editing on all editable columns (Winner and Notes)
-            row_count = table.rowCount()
-            for row in range(row_count):
-                for col in [2, 3]:
-                    item = table.item(row, col)
-                    if item is not None:
-                        item.setFlags(item.flags() & ~Qt.ItemFlag.ItemIsEditable)
-
-                widget = table.cellWidget(row, 2)
-                if isinstance(widget, QComboBox):
-                    widget.setEnabled(False)
-
-            # Lock the round
-            round.locked = True
-
-            # Update the scores for every player
-            for matchup in round.matchups:
-                matchup.player1.matches.append(matchup)
-                matchup.player1.score += matchup.score_player1
-                if matchup.player2:
-                    matchup.player2.matches.append(matchup)
-                    matchup.player2.score += matchup.score_player2
-            
-            # TODO: Can probably be a lot better
-            # Update the win percentrage for every player
-            for player in self.players:
-                total_wins = 0
-                for match in player.matches:
-                    if match.winner == player:
-                        total_wins += 1
-                player.winpercentage = total_wins / len(player.matches) * 100
-
-            # Update resistance
-            for player in self.players:
-                owps = []
-                print(player)
-                for played_round in self.rounds:
-                    for match in played_round.matchups:
-                        print(match)
-                        # Find the player's match
-                        if player == match.player1 or player == match.player2:
-                            opponent = match.player1 if match.player1 != player else match.player2
-                            if opponent:
-                                owps.append(max(opponent.winpercentage, 25))
-                            else:
-                                # This was BYE round so ignore it
-                                pass
-                            break
-
-                if owps:
-                    player.resistance = sum(owps) / len(owps)
-                else:
-                    # no matches found so first round was a BYE
-                    player.resistance = 25
-
-            self.update_player_table()
-
-            print("Round locked! Editing disabled.")
-
-
-    def update_player_table(self):
-        row_count = self.ui.playersTableWidget.rowCount()
-
-        for row in range(row_count):
-            # update score, resistance, winpercentage and matches
-            name = self.ui.playersTableWidget.item(row, 1).text()
-            player = get_player_by_name(self.players, name)
-            print(player)
-            self.ui.playersTableWidget.setItem(row, 2, QTableWidgetItem(f"{player.score}"))
-            self.ui.playersTableWidget.setItem(row, 3, QTableWidgetItem(f"{player.resistance}"))
-            self.ui.playersTableWidget.setItem(row, 4, QTableWidgetItem(f"{player.winpercentage}"))
 
 
     def on_generate_final_bracket_clicked(self):
@@ -536,24 +600,23 @@ class MainWindow(QtWidgets.QMainWindow):
 
         # Clean text with padding and slightly larger font
         choice_box.setText(
-            "<div style='padding: 6px; font-size: 11pt;'>"
+            "<div style='padding: 32px; font-size: 11pt;'>"
             "How would you like to select players for the final bracket?"
             "</div>"
         )
 
         choice_box.setIcon(QMessageBox.Icon.NoIcon)
-        choice_box.setMinimumWidth(420)  # Slightly wider than before
 
         # Add buttons
         top_x_button = choice_box.addButton("Top X Players", QMessageBox.ButtonRole.AcceptRole)
         score_cut_button = choice_box.addButton("Score Threshold", QMessageBox.ButtonRole.AcceptRole)
         cancel_button = choice_box.addButton("Cancel", QMessageBox.ButtonRole.RejectRole)
 
-        # Style the buttons with normal font (no bold), slightly larger
-        button_style = "padding: 6px 12px; font-size: 10.5pt;"
-        top_x_button.setStyleSheet(button_style)
-        score_cut_button.setStyleSheet(button_style)
-        cancel_button.setStyleSheet(button_style)
+        # Style the buttons
+        button_style = "padding: 6px 12px; font-size: 9pt;"
+        for button in (top_x_button,score_cut_button,cancel_button):
+            button.setStyleSheet(button_style)
+            button.setMinimumWidth(140)
 
         # Show the message box
         choice_box.exec()
@@ -564,8 +627,6 @@ class MainWindow(QtWidgets.QMainWindow):
             self.show_top_x_input()
         elif selected_button == score_cut_button:
             self.show_score_threshold_input()
-
-
 
     def show_top_x_input(self):
         dialog = QDialog(self)
@@ -592,9 +653,11 @@ class MainWindow(QtWidgets.QMainWindow):
                 if num > len(self.players):
                     QMessageBox.warning(self, "Invalid Input", "More players selected than are listed.")
                     return
-                sorted_players = sorted(self.players, key=lambda p: (p.score, p.resistance), reverse=True)
-                print(sorted_players)
-                selected_players = sorted_players[:num]
+
+                player_info_list = calculate_players_stats(self.players, self.rounds)
+                sorted_players_info = sorted(player_info_list, key=lambda p: (p.score, p.resistance), reverse=True)
+                print(sorted_players_info)
+                selected_players = sorted_players_info[:num]
                 if len(selected_players) == 0:
                     raise ValueError()
                 round1_matches = create_bracket(selected_players)
@@ -632,15 +695,16 @@ class MainWindow(QtWidgets.QMainWindow):
             try:
                 wins = int(text)
                 threshold_score = wins  # Adjust if you use a different score metric
-                selected_players = [p for p in self.players if p.score >= threshold_score]
-                if len(selected_players) < 2:
+                player_info_list = calculate_players_stats(self.players, self.rounds)
+                selected_players_info = [p for p in player_info_list if p.score >= threshold_score]
+                if len(selected_players_info) < 2:
                     QMessageBox.warning(self, "Invalid Input", "Not enough players have a high enough score.")
                     return
-                sorted_players = sorted(selected_players, key=lambda p: (p.score, p.resistance), reverse=True)
-                round1_matches = create_bracket(sorted_players)
+                sorted_players_info = sorted(selected_players_info, key=lambda p: (p.score, p.resistance), reverse=True)
+                round1_matches = create_bracket(sorted_players_info)
                 # TODO: Future work. Create full interactive bracket page
                 # bracket = build_full_bracket_from_first_round(round1_matches)
-                self.show_classic_bracket(round1_matches)
+                # self.show_classic_bracket(round1_matches)
                 dialog.accept()
             except ValueError:
                 QMessageBox.warning(self, "Invalid Format", "Enter valid numbers only.")
@@ -650,6 +714,8 @@ class MainWindow(QtWidgets.QMainWindow):
 
         dialog.exec()
 
+
+
     def show_classic_bracket(self, matchups: list[Matchup]):
         """
         Displays a single-round bracket as a table for copy-pasting to Excel.
@@ -657,8 +723,17 @@ class MainWindow(QtWidgets.QMainWindow):
 
         # Create the table
         table = QTableWidget()
-        table.setColumnCount(6)
-        table.setHorizontalHeaderLabels(["P1", "P2", "Winner", "Notes", "P1Score", "P2Score"])
+
+        table.setContextMenuPolicy(Qt.ContextMenuPolicy.ActionsContextMenu)
+
+        copy_action = QAction("Copy", table)
+        copy_action.setShortcut(QKeySequence.StandardKey.Copy)
+        copy_action.triggered.connect(lambda: self.copy_selection_to_clipboard(table))
+
+        table.addAction(copy_action)
+
+        table.setColumnCount(2)
+        table.setHorizontalHeaderLabels(["P1", "P2"])
         table.setSortingEnabled(False)
 
         # Enable multi-cell selection
@@ -667,51 +742,15 @@ class MainWindow(QtWidgets.QMainWindow):
 
         for idx, matchup in enumerate(matchups):
             table.insertRow(idx)
-
-            p1 = matchup.player1
-            p2 = matchup.player2 if matchup.player2 else Player("-")
-            winner = Player("")
-
-            if p2.name == "-":
-                winner = p1
-                matchup.score_player1 = 1.0
-
-            table.setItem(idx, 0, QTableWidgetItem(p1.name))
-            table.setItem(idx, 1, QTableWidgetItem(p2.name))
-
-            combo = QComboBox()
-            combo.setEditable(True)
-            combo.lineEdit().setReadOnly(True)
-            combo.lineEdit().setPlaceholderText("Select winner...")
-
-            combo.addItem(p1.name)
-            if p2.name != "-":
-                combo.addItem(p2.name)
-
-            if winner.name:
-                combo.setCurrentText(winner.name)
-            else:
-                combo.setCurrentIndex(-1)
-
-            combo.currentTextChanged.connect(
-                lambda name, row=idx: self.on_winner_changed(row, name, table, matchups)
-            )
-            table.setCellWidget(idx, 2, combo)
-
-            table.setItem(idx, 3, QTableWidgetItem(str(matchup.notes)))
-            table.setItem(idx, 4, QTableWidgetItem(str(matchup.score_player1)))
-            table.setItem(idx, 5, QTableWidgetItem(str(matchup.score_player2)))
-
-        table.cellChanged.connect(
-            lambda row, col, t=table, m=matchups: self.on_cell_changed(t, m, row, col)
-        )
+            table.setItem(idx, 0, QTableWidgetItem(matchup.player1))
+            table.setItem(idx, 1, QTableWidgetItem(matchup.player2))
 
         # Container widget
         container = QWidget()
         layout = QVBoxLayout(container)
 
         # Add a copy button for first 2 columns
-        copy_button = QPushButton("Copy P1 & P2 columns to clipboard")
+        copy_button = QPushButton("Copy player columns to clipboard (for exporting to Excel)")
         copy_button.clicked.connect(lambda: self.copy_first_two_columns(table))
         layout.addWidget(copy_button)
 
@@ -720,6 +759,28 @@ class MainWindow(QtWidgets.QMainWindow):
 
         self.ui.tabWidget.addTab(container, "Final Bracket")
         self.ui.settingsMessage.setText("Bracket matchups displayed.")
+
+
+    def copy_selection_to_clipboard(self, table: QTableWidget):
+        selection = table.selectedRanges()
+        if not selection:
+            return
+
+        selected_range = selection[0]  # contiguous block
+        rows = range(selected_range.topRow(), selected_range.bottomRow() + 1)
+        cols = range(selected_range.leftColumn(), selected_range.rightColumn() + 1)
+
+        lines = []
+        for row in rows:
+            line = []
+            for col in cols:
+                item = table.item(row, col)
+                line.append(item.text() if item else "")
+            lines.append("\t".join(line))
+
+        text = "\n".join(lines)
+        QApplication.clipboard().setText(text)
+
 
     def copy_first_two_columns(self, table: QTableWidget):
         rows = table.rowCount()
@@ -732,8 +793,81 @@ class MainWindow(QtWidgets.QMainWindow):
         self.ui.settingsMessage.setText("Copied first two columns to clipboard.")
 
 
+class StartupDialog(QDialog):
+    def __init__(self):
+        super().__init__()
+        self.setWindowTitle("Swiss Bracket Maker")
+        self.choice = None
+
+        layout = QVBoxLayout(self)
+        label = QLabel("How do you want to launch?")
+        label.setAlignment(Qt.AlignmentFlag.AlignCenter)
+        layout.addWidget(label)
+
+        layout.addStretch(1)
+
+        label.setStyleSheet("""
+            QLabel {
+                font-size: 18px;
+                font-weight: 600;
+                padding: 20px;
+            }
+        """)
+        label.setMinimumHeight(100)
+
+        button_layout = QHBoxLayout()
+        # Spacing between the 2 buttons
+        button_layout.setSpacing(16)
+        # Margin around the horizontal layout
+        button_layout.setContentsMargins(16, 0, 16, 16)
+        button_new = QPushButton("New")
+        button_prev = QPushButton("Open tournament file")
+        button_new.setMinimumSize(160, 36)
+        button_prev.setMinimumSize(160, 36)
+        button_layout.addWidget(button_new)
+        button_layout.addWidget(button_prev)
+
+        button_new.clicked.connect(lambda: self.select("new"))
+        button_prev.clicked.connect(lambda: self.select("prev"))
+
+        layout.addLayout(button_layout)
+
+    def select(self, value):
+        if value == "prev":
+            self.choice = self.read_input_file()
+
+        self.accept()
+
+    def read_input_file(self) -> dict:
+        file_name, _ = QFileDialog.getOpenFileName(
+            self,
+            "Load Tournament Data",
+            "",
+            "JSON Files (*.json)"
+        )
+        if not file_name:
+            return
+        # Parse the file
+        try:
+            with open(file_name, "r", encoding="utf-8") as f:
+                return json.load(f)
+        except Exception as e:
+            QMessageBox.critical(self, "Import Failed", f"Could not load file:\n{e}")
+            return None, None
+
 if __name__ == '__main__':
+
     app = QtWidgets.QApplication([])
-    window = MainWindow()
+
+    dialog = StartupDialog()
+    dialog.exec()
+
+    # Launch main window based on choice
+    print(dialog.choice)
+    window = MainWindow(dialog.choice)
+
+
+
+    # window = MainWindow()
     window.show()
-    app.exec_()
+    app.exec()
